requirements.txt
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.0.0
wandb>=0.17.0
